windows: Refactor code for injecting drivers into a library for easier shared use
(cherry picked from commit 8b651a1235f7c3df615e9cecb137a2187fd9629f)
License: Apache-2.0
